rename metro settings
