+++ metroSettings/src/main/AndroidManifest.xml
