+++ metroSettings/src/main/res/drawable/ic_baseline_rate_review_24.xml
