+++ metroSettings/src/main/res/drawable/ic_baseline_share_24.xml
